gh-178 use Command Objects for data binding and change their package to rest.transport
