+++ mdm-plugin-profile/grails-app/views/profile/many.gson
@@ -1,5 +1,5 @@
 import uk.ac.ox.softeng.maurodatamapper.gorm.PaginatedResultList
-import uk.ac.ox.softend.maurodatamapper.profile.databinding.ProfileProvided
+import uk.ac.ox.softeng.maurodatamapper.profile.rest.transport.ProfileProvided
 
 model {
     Iterable<ProfileProvided> profileProvidedList
@@ -7,6 +7,6 @@ model {
 
 json {
     count profileProvidedList instanceof PaginatedResultList ? ((PaginatedResultList) profileProvidedList).getTotalCount() : profileProvidedList?.size() ?: 0
-    profilesProvided tmpl.'/profile/profileProvided'(profileProvidedList ?: [])
+    profilesProvided tmpl.'/profileProvided/profileProvided'(profileProvidedList ?: [])
 }
