+++ mdm-plugin-profile/grails-app/views/profileProvided/_profileProvided.gson
@@ -1,4 +1,4 @@
-import uk.ac.ox.softend.maurodatamapper.profile.databinding.ProfileProvided
+import uk.ac.ox.softeng.maurodatamapper.profile.rest.transport.ProfileProvided
 import uk.ac.ox.softeng.maurodatamapper.profile.object.Profile
 
 model {
